:zap: extracted fuzzyCompare into utils, refactor
